+++ src/test/java/ca/ulaval/glo2003/bookings/domain/BookingTotalCalculatorTest.java
@@ -20,7 +20,7 @@ class BookingTotalCalculatorTest {
 
   private static BookingTotalCalculator bookingTotalCalculator;
 
-  private static Price total = new Price(BigDecimal.valueOf(100));
+  private static Price total = new Price(100);
   private static Price pricePerNight = mock(Price.class);
   private static Bed bed = mock(Bed.class);
   private static Packages bedPackage = createPackageName();
@@ -53,9 +53,9 @@ class BookingTotalCalculatorTest {
 
   private static Stream<Arguments> provideConditionsToCalculateTotal() {
     return Stream.of(
-        Arguments.of(1, new Price(BigDecimal.valueOf(100))),
-        Arguments.of(3, new Price(BigDecimal.valueOf(95))),
-        Arguments.of(10, new Price(BigDecimal.valueOf(90))),
-        Arguments.of(30, new Price(BigDecimal.valueOf(75))));
+        Arguments.of(1, new Price(100)),
+        Arguments.of(3, new Price(95)),
+        Arguments.of(10, new Price(90)),
+        Arguments.of(30, new Price(75)));
   }
 }
